+++ packages/runtime-utils/src/index.ts
@@ -10,7 +10,9 @@ import {
 
 export { getErrorMessage } from "./errors";
 export * from "./guards";
+export * from "./logger";
 export * from "./objects";
+export * from "./sensitive-key";
 
 export type JsonValueValidationOptions = Readonly<{
 	maxDepth?: number;